rename old tokenInvocation => getTokenInvocationArgs
