+++ @stellar/typescript-wallet-sdk-soroban/src/Helpers/index.ts
@@ -1,5 +1,5 @@
 export * from "./formatTokenAmount";
 export * from "./getInvocationDetails";
+export * from "./getTokenInvocationArgs";
 export * from "./parseTokenAmount";
 export * from "./scValByType";
-export * from "./tokenInvocation";
